Restructure Rust source files
Instead of everything being in cpio_cache, things now live in more
appropriate places.
