+++ src/lib.rs
@@ -1,5 +1,8 @@
 mod cpio;
 pub mod cpio_cache;
+mod cpio_lru_cache;
+mod cpio_maker;
+mod error;
 mod files;
 mod nix;
 mod opened_cpio;
